+++ src/AckModule.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { FormsModule }   from '@angular/forms';
-import{ HttpModule } from '@angular/http';
+import{ HttpClientModule } from '@angular/common/http';
 //import { Ng2PageScrollModule } from 'ng2-page-scroll';
 
 import { providers } from "./providers";
@@ -15,13 +15,13 @@ const declarations = [...components, ...pipes]
   imports:[
     CommonModule
     ,FormsModule
-    ,HttpModule
+    ,HttpClientModule
     //,AckOffline
     //,BrowserAnimationsModule
     //,Ng2PageScrollModule.forRoot()
   ]
   ,declarations: declarations
   ,providers:providers,
-  exports:[FormsModule, HttpModule, ...declarations]
+  exports:[FormsModule, HttpClientModule, ...declarations]
 }) export class AckModule {}